Add patient info CANCER_TYPE* if in all samples
If all samples have the same CANCER_TYPE or CANCER_TYPE_DETAILED, add them to
the patient info at the top of the page instead of showing them per sample.
This removes the code of only showing CANCER_TYPE and CANCER_TYPE_DETAILED if
the sample is primary such that attributes are still displayed if there are
only metastases for a patient and no primary.