Draw sample labels on timeline
The clinical-timeline has been updated to allow hooks after drawing the
timeline, that way some of the elements on the timeline can be adjusted every
time the timeline display is updated, either due to hiding of a track or
splitting etc. Patients with one sample get a label now as well.