Add Travis CI screenshot end-to-end tests
Add simple test to Travis CI that uses phantomjs to make a screenshot of both
the study view and the patient view and compares them against the screenshots
stored in the repo. If they differ, upload the test screenshot on clbin.com for
further inspection. The backend used is the amazon public db that is also used
by heroku.