Integration of CIVIC database
The integration of CIVIC follows the approach taken by integration of the
change_hotspot list. Static TSV files are added as resources to the WAR
file, with the option to use files from the file system instead through
configuration.

The logo of CIVIC is (c) The Genome Institute, inquiry about using their
logo can be found here:

- https://github.com/genome/civic-client/issues/398