Added new persistence and service modules
This module is using the new model module. Converted ProfileDataMapper.getMutationData to the new
modules as an example (didn't delete the original one in business module).
Didn't add the new model package to typeAliasesPackage because there are classes that has the same name in
org.mskcc.cbio.portal.model package, so we need to use explicit model class names in mapper xmls.
Wanted to turn mapUnderscoreToCamelCase setting on but it breaks the current code. We can make those changes
after we delete old models/modules.