Add online image compare URL for failed screenshot
For failing screenshots: output a rawgit URL where one can compare the images
online. The image compare viewer is in test/end-to-end/image-compare. The image
URLs as well as labels for the images can be supplied as arguments to the URL
allowing one to compare any combination of images. Updated the CONTRIBUTING
docs to reflect the change. The image compare viewer also includes some
documentation of how to handle a difference in the image.

Signed-off-by: Ino de Bruijn <ino@ino.pm>