Make study summary top bar buttons behave as buttons
Because the buttons to view or download selected samples are not
implemented using true `<button>` elements, they did not previously
appear to assistive technologies as buttons, and they would not be
operable by other means than a mouse. This patch fixes that by
adding the appropriate ARIA role, adding them to the control element
navigation order (tabindex), and responding to space bar.