Beginning session filter support
Introduce to the QuerySession the notion of a session filter,
which is a data filter installed into the QuerySession object
by a caller - the motivating use case is the oncoprint declaring
that non-recurrent mutations are to be treated as nonexistent,
and wanting other functions (altered samples, etc) to reflect that.

This commit has beginning capabilities for this and a corresponding
set of methods (e.g. getSessionFilteredWebServiceGenomicEventData)
that deal with the session filters.