Add tab for MDACC Heatmap viewer on study page
This adds a new tab to the study view page.  Clicking on the tab
displays an embedded "MDACC Clustered Heatmap Compendium" page, that
shows only heatmaps that were indexed as belonging with the study being
viewed.  If the user clicks on a heatmap, The NG-CHM viewer is displayed
within that embedded iframe.  The approach used (the embedded iframe)
follows the approach for the existing embedded Digital Slide Archive
viewer.

This is based on the 7/5/16 phone discussion with Niki Schultz, Ethan
Cerami, John Weinstein, and myself, when we demoed a patient view
version of the same.

The heatmap index mentioned above is queried via a web service running
at MD Anderson, whose URL is set in the property
mdacc.heatmap.study.meta.url in portal.properties.EXAMPLE.  A blank or
missing url property causes the query to be skipped and disables the
entire functionality.  For development purposes, the test study in the
"Seed Database" with id 'study_es_0' is mapped to the same heatmaps
that we have for study id 'brca_tcga_pub' that contains similar
samples.

Signed-off-by: Chris Wakefield <cwakefield@mdanderson.org>