Remove db validation for clinical attributes
Clinical attributes used to share a single namespace across all
studies, meaning that once the meaning of an attribute name was
defined (with a display name, mouseover description, data type,
sample or patient level and priority), all subsequent studies would
have to adhere to that. Commit b6dcfc082116293178b0cd762b0be7aba3152d51
made clinical attributes study-specific. Because of that, validating
whether the definitions of attributes in a study are consistent
with those configured in the portal is now no longer applicable.