Add tab for MDACC Heatmap viewer on patient view
This adds a new tab to the patient view page.  Clicking on the tab
displays an embedded "MDACC Clustered Heatmap Compendium" page, that
shows only heatmaps that were indexed as including the sample id being
viewed.  If the user clicks on a heatmap, The NG-CHM viewer is displayed
within that embedded iframe.  The approach used (the embedded iframe)
follows the approach for the existing embedded Digital Slide Archive
viewer.

This is based on my previous pull request #1709

The heatmap index mentioned above is queried via a web service running
at MD Anderson, whose URL is set in the property
mdacc.heatmap.meta.url in portal.properties.EXAMPLE.  A blank or
missing url property causes the query to be skipped and disables the
entire functionality.

Signed-off-by: Chris Wakefield <cwakefield@mdanderson.org>