Hide unsequenced in oncoprint, and make view menu say Show unaltered samples/patients instead of columns
Also fix two bugs that presented themselves
(1) When vertex column is undefined, as when the column is hidden in the oncoprint,
we were passing in undefined into the webGL shader. I've modified it to pass in the dummy
off-screen column -1000.

(2) When changing from sample to patient data, the ids which are registered to be zoomed
to fit, sample uids, are no longer visible in the oncoprint, and so the math that
determines it ends up doing Math.max(-infinity, undefined) which gives NaN. These
NaNs were breaking the oncoprint. To fix the bug, I added a branch to that code
to filter out undefined values from the calculation.

Signed-off-by: Abeshouse, Adam A./Sloan Kettering Institute <abeshoua@mskcc.org>