Adjust data types that support a gene panel field
The Python code for parsing and validating metadata files in a study
are aware of which fields will be recognised for each data type,
but as some of the data types that do support the gene panel field
were not listed, these yielded a misleading warning. This commit
suppresses the warning for these data types, and removes one
erroneously documented data type from File-Formats.md.