Limit `getGeneByHugoSymbol()` to 1 result in `GenePanelMapper.xml`
The logic in `DaoGeneOptimized` returns the first Gene object if multiple records are returned when querying by a hugo symbol. This PR extends this logic to the `GenePanelMapper.xml`.

The `getGeneByHugoSymbol()` method expects only one result but there are cases where a hugo symbol is not unique and is actually associated with multiple entrez gene ids. In its current state, a `TooManyResultsException` will be thrown if a query returns more than one record for a given hugo symbol.

Signed-off-by: Angelica Ochoa <aochoa4230@gmail.com>