Remove special case for SAML institute logo URL
The template logic for the header bar contained a special case in which
the institute logo was be prepended with a slash when using SAML
authentication, which caused issues when no logo was applicable
(signalled by an empty string). The logo seems to display just fine
without the special case, and removing it both solves the issue and
gets rid of some non-templating logic in a JSP scriptlet.