Fix broken institute logo in older Tomcat versions
The JSP template for the header bar used an EL expression which called a
static Java method. This turned out not to work in Tomcat 7, causing the
template to insert a broken <img> tag no matter whether a logo was
enabled or not. The native JSP way to solve this would be to define the
static method as a function in a taglib, but as this is just a single
function call I decided to keep it short and evaluate a Spring SpEL
expression instead.