Remove threshold on calculating significant mutation genes
The current threshold creates inconsistency across pages.
For instance, RYR2 in luad_tcga_pub is highly mutated, but since the algorithm is designed as # mutations / gene length which causes RYR2 ranked really low will eventually be trimmed off from the list. But on the other hand, OncoPrint is showing all genes.
This solves #3305

Signed-off-by: Hongxin Zhang <jackson.zhang.828@gmail.com>