Fix #3330 Make query button work for study view
This fixes the query button on study view for multiple studies. Also fixes
issues with passing of case lists and studies to the backend. We now no longer
create a virtual cohort on the frontend but instead just pass the relevant
studyIds and caseLists. Works both for when the session service is enabled or
disabled.