Fix unique keys in gene set data API response
In order to use the endpoint to make heatmaps of gene set scores, the
unique keys should uniquely identify samples and patients across
studies, where they previously contained nulls. In addition, Ersin
suggested leaving the IDs of the patients and studies in the JSON as
well, since we're reading them anyway.