Pass sessionServiceIsEnabled to frontend
Pass variable to indicate that the session service is enabled. My main purpose
is for testing so one can disable session service tests, but I am sure there
are plenty of other situations where it's useful to know this. Could for
instance save a request on the home page when checking if there are any virtual
studies.