Fix the validateStudies flush-to-console crash
The logging framework used by the validation script does not expect to
run in the same Python interpreter process as another run, and doing
this anyway caused problems particularly when running the multi-study
validator on studies with more than a million (possible) errors. Running
the validation script in a subprocess fixes.

One small complication of not having access to the same logger was that
the study header had to be explicitly flushed to the log file before
running the validation script, so that it doesn't stay in a buffer until
a later line.