Make scripts find their JAR relative to themselves
When using the PORTAL_HOME environment variable to set the runtime
config folder to use, it may not always be a compiled source code
directory. A more robust way for a script to find another piece of code
in the same project, not uncommon in the Python world, is relative to
the script itself.