Document deployment before data loading in manual
The previous ordering stems from the time when cBioPortal issued a
confusing error message when started without any studies in the
database; the current handles this cleanly.

This creates a more natural pattern leading from databse configuration
to webserver deployment to data loading.

It may also open up the future possibility of loading the initial study
without disabling the parts of the pipeline that depend on webservices.