Make `authenticate=` a required JVM argument
It cannot be configured at runtime if it's included in portal.properties
at buildtime, as Maven overwrites web.xml and makes it nonconfigurable.
Removing it from portal.properties.EXAMPLE will at least fix that for
installations that do not touch the buildtime portal.properties file.

In addition, I don't know how to set the Spring profile to be used in
applicationContext-security.xml based on portal.properties, as I believe
web.xml runs *before* GlobalProperties.java.

To fix the runtime configuration setup documented in the deployment
manual now, make authenticate= a JVM system property instead. If not
configured at build-time as before, it must be supplied in
CATALINA_OPTS.