Suggest explicit review requesting guidelines
Aiming to avoid the situation in which it's unclear whether someone is
still writing  a review or has decided not to prioritise it since they
wouldn't find much that other reviewers haven't already found.

Also documenting the practice of notifying reviewers who know more about
use-cases and architectural decisions related to the logic/docs being
changed.