Disable Pool Prepared Statement JDBC
We are using -Ddbconnector=jdbc on AWS in production. Before the dbcp was only
being used for the importer. It seems that enabling this when serving the
webapp results in a huge number of queries being stored in memory. We might
eventually want to make this some property that one can only enable when using
the importer.