Travis fixes related to oraclejdk8 deprecation
The oraclejdk8 we were using was deprecated. Switching to openjdk8
requires a new Ubuntu image, which required several changes.

- add mysql service on travis
- use docker to use python 3.4 on Travis. The new Ubuntu image has 3.5
by default, but the assert_called method in the tests does not exist
there. They do exist in version 3.6 and 3.7 but those currently aren't
easily installed through apt-get i.e. one needs to add a non-default
repo to download it
- quiet maven when building Travis for e2e compilation. Apparently there
is a limit on the Travis log size for each build. When building the
portal for e2e tests the log limit on Travis gets reached resulting in
an error.