Fixed validation issue with different geneset_def_version
The current validation process of cBioPortal studies does not throw an
error when the property geneset_def_version is different in the study
from the one in the database. This leads to an error throwing when
cBioPortal tries to load the study.
Fixed it by first adding the geneset version from database to the API
and also to PortalInfo, then retrieving it in the validateData.py script
and comparing it to geneset_def_version in the study.