Publish tags to docker hub
Fix #7442. Publish docker images for each tag on github

Note that we can't reuse the existing master image as proposed in the issue, because the version in the pom is determined on the fly from git i.e. the latest master version shows a different version in the api/info response than a proper tagged version even though the code might be completely the same otherwise.