Fix bug in data access token download.
Argument for 'allowRemovalOfExistingTokens' has been disconnected or
removed in some way.

When a user attempts to download a token from the
login dropdown menu, if the option to revoke existing tokens was
enabled then the user would be presented with an error if they had an
existing token instead of being issued a brand new one.

We have decided to simply remove this functionality and always issue a
new token on request.

Signed-off-by: Angelica Ochoa <15623749+ao508@users.noreply.github.com>