Optimize mutation count by keyword calculation
Current implementation of this calculation is extremely time consuming
for studies with large and/or multiple MAFs.

With the changes in this PR the runtime for the calculation is reduced
from greater than 60 minutes to approximately 18 minutes for the
mskimpact study.

In short the changes in this PR implement the use of the bulk loader as
well as a step that pre-calculates the gene counts and mutation keyword
counts for the current study being imported before creating the insert
statements with the bulk loader and executing the import.

Signed-off-by: Angelica Ochoa <15623749+ao508@users.noreply.github.com>