Custom Redis Cache
This PR provides a custom implementation of Spring's cache, backed
by Redis. We decided to do this because we found that the
CachingProvider that Spring uses for Redis did not work well with
our set up. Specifically, it did not give us enough autonomy
over expiry, and did not allow us to compress objects.

This PR provides:
- A custom redis cache manager that compresses values and sets ttls
- A revamped key generator
- A ton of tests for the cache
- Caching of a few controller methods
- Implementation of cache stats for Redisson client.