fix #8614 - caching the result of getSampleListSampleIds for a single sampleListId
fix #8614 - caching the result of getSampleListSampleIds for a single sampleListId; Instead of calling getSampleListSampleIds with a long list of sample List Ids, calling getSampleListSampleIds with a single SampleListId so that the results can be cached. This will reduce load on the db server significantly.

fix #8594 - Cleanup and refactoring. Removing the instance object.