Fix validation of resource data for patient
For some reason this gives an error when using a single patient with
associated resource data. It is e.g. checking whether a particular
patient has samples, but in the case where a resource is attached to a
patient directly it does not matter it the patient has samples at all.
The variable `RESOURCE_PATIENTS_WITH_SAMPLES` is not defined. Not
entirely sure how to best check for that case, so I just removed the
validation altogether